moved all non-route methods to 3rd party libs
