+++ wsgi/gameio.py
@@ -1,0 +1,157 @@
+"""
+INPUT / OUTPUT (SMS, signup auth, eail, etc.)
+"""
+
+#This method is 'dumb'. All it does is accept data, build a payload, and schedule a job. If the job is queued successfully, it returns a task/job id. It doesen't know it's sending an SMS!!!
+def sendSMS(f,t,m,u,d,st):
+    # f - from
+    # t - to
+    # m - msg
+    # u - url
+    # d - delay
+    # st - absolute send time
+    #print "####"
+
+    #worker = IronWorker()
+    worker = IronWorker(project_id=environ['IID2'], token=environ['ITOKEN2'])
+    #worker = IronWorker(project_id=ironInfo['project_id'], token=ironInfo['token'])
+    #print worker
+
+    task = Task(code_name="smsworker", scheduled=True)
+    #print task
+
+    task.payload = {"keys": {"auth": environ['TSID'], "token": environ['TTOKEN']}, "fnum": f, "tnum": t, "msg": m, "url": u}
+    #print task.payload
+
+    # scheduling conditions
+    if d is not None:
+        task.delay = d
+        #print "sending SMS after", d, "second delay"
+        log.info('sending sms after %s second delay' % (d) )
+    elif st is not None:
+        task.start_at = st # desired `send @ playertime` converted to servertime
+        #print "sending SMS at:", st
+        log.info('sending sms at %s' % (st))
+    else:
+        task.delay = 0
+        #print "sending SMS right away"
+        log.info('sending sms right away')
+
+    # now queue the damn thing & get a response.
+    response = worker.queue(task)
+    #print response
+    return response.id
+
+def sendEmail(fe, fn, te, tn, sub, txt, html, d, st):
+    # dom - MG domain
+    # key - MG api key
+    # fe - npc email
+    # fn - npc display_name
+    # te - player email
+    # tn - player name
+    # sub - subject line
+    # txt - text version of email
+    # html - html version of email
+
+    #worker = IronWorker()
+    worker = IronWorker(project_id=environ['IID2'], token=environ['ITOKEN2'])
+
+    task = Task(code_name="emailworker", scheduled=True)
+    task.payload = {
+        "dom": environ['MGDOM'],
+        "key": environ['MGKEY'],
+        "fe": fe,
+        "fn": fn,
+        "te": te,
+        "tn": tn,
+        "sub": sub,
+        "txt": txt,
+        "html": html
+        }
+
+    # scheduling conditions
+    if d is not None:
+        task.delay = d
+        #print "sending email after", d, "second delay"
+        log.info('sending email after %s second delay' % (d) )
+    elif st is not None:
+        task.start_at = st # desired `send @ playertime` converted to servertime
+        #print "sending email at:", st
+        log.info('sending email at %s' % (st))
+    else:
+        task.delay = 0
+        #print "sending email right away"
+        log.info('sending email right away')
+
+    # now queue the damn thing & get a response.
+    response = worker.queue(task)
+    #print response
+    return response.id
+
+def signupSMSauth(tel,auth):
+    # lookup admin NPC # for the user's country (this of course, assumes we have one)
+    fnum = getNPC({ "country": getCountryCode(tel) }, 'admin')['tel']
+    #msg = "code: " + str(auth) +" "+ u"\U0001F6A8"
+    msg = "--Mercury Group HR system--\nThank you for your application.\nYour 4 digit identification code is: " + str(auth)
+
+    # send auth SMS
+    workerStatus = sendSMS(fnum,tel,msg,None,0,None) #no media, 0s delay, no sendTime
+    #print workerStatus
+
+    if workerStatus is not None:
+        #print "worker id", workerStatus
+        log.info('worker id %s' % (workerStatus))
+        return True
+    else:
+        #print "worker error - probably"
+        log.debug('worker error ~ probably')
+        return False
+
+# THIS METHOD IS NOT COMPLETE
+def sendErrorSMS(player):
+    # send random error phrase from list to user
+    err = random.choice(errlist)
+    print "error msg: " + err
+    log.debug('error SMS msg: %s' % (err))
+
+    gs = getGameStateData(player['gstate'])
+    npc = getNPC(player, gs['prompt']['npc'])
+
+    sendSMS(npc['tel'], player['tel'], err, None, 14, None)
+
+# SMS auth - store token
+def newAuth(a):
+    db = create_engine(environ['OPENSHIFT_MYSQL_DB_URL'] + environ['OPENSHIFT_APP_NAME'], convert_unicode=True, echo=False)
+    #db = create_engine(Mdb, convert_unicode=True, echo=False)
+    md = MetaData(bind=db)
+    table = Table('tokenAuth', md, autoload=True)
+    con = db.connect()
+
+    now = datetime.datetime.now()
+    d = now.strftime('%Y-%m-%d %H:%M:%S')
+
+    x = con.execute( table.insert(), auth=a, date=d)
+    uid = x.inserted_primary_key[0]
+    con.close()
+    return uid
+
+# SMS auth - return auth based on token
+def checkAuth(uid):
+    db = create_engine(environ['OPENSHIFT_MYSQL_DB_URL'] + environ['OPENSHIFT_APP_NAME'], convert_unicode=True, echo=False)
+    #db = create_engine(Mdb, convert_unicode=True, echo=False)
+    md = MetaData(bind=db)
+    table = Table('tokenAuth', md, autoload=True)
+    con = db.connect()
+    x = con.execute( table.select(table.c.auth).where(table.c.uid == uid) )
+    a = x.fetchone().get('auth', None)
+    con.close()
+    return a
+
+def getCountryCode(tel):
+    tel = normalizeTel(tel)
+    lookup = TwilioLookupsClient(environ['TSID'], environ['TTOKEN'])
+    return lookup.phone_numbers.get(tel).country_code
+
+def normalizeTel(tel):
+    nTel = re.sub(r'[^a-zA-Z0-9\+]','', tel)
+    return nTel
